requirements.txt
async-generator==1.10
attrs==22.2.0
blinker==1.5
Bottleneck==1.3.4
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==39.0.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
kaitaistruct==0.10
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numexpr==2.8.1
numpy==1.22.3
openpyxl==3.0.9
outcome==1.2.0
packaging==21.3
pandas==1.4.2
pyasn1==0.4.8
pycparser==2.21
pydivert==2.1.0
pyOpenSSL==23.0.0
pyparsing==3.0.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
requests==2.28.1
selenium==4.7.2
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
webdriver-manager==3.8.5
wincertstore==0.2
wsproto==1.2.0
zstandard==0.19.0
